tests: allows cluster to suppress process crashes
Adds an option to @cluster to allow for missing redpanda processes,
as the new chaos test can cause processes to die which is expected.

Also changes some logging from warning to info, and allows override
of config when using helper functions.

the random action injected by the context is now applied before
sleep, so that at least one action runs before sleep finishes.

Also refactor the tests so that the base class is used for SI
e2e tests, and uses Event instead of boolean to stop thread.

when nodes are killed during tests, validations should not run
on killed nodes, this commit syncs redpanda.started to remove
killed nodes.

also, the default replication factor is changed to num brokers,
so that if the group leader is killed the system can still recover.