tests: allow action injector to reverse actions
adds code to allow action injector to optionally reverse last
action when applying the current action. this allows the code to
avoid pile-up of killed nodes etc.

also adds action log which can be used in assertions, and adds
node recommission helper to admin wrapper.

as nodes are killed during tests some errors need to be ignored
which are found in CHAOS_LOG_ALLOW_LIST